Released Confirm-Address 1.3.4
Supported WebExtension.
Be aware of installation: it's irreversible.  Once you install Confirm-Address 1.3.4, you won't be able it back to 1.3.3 or older.
"unittest/" folder is excluded from the file sent to AMO, otherwise it would be warned as it contains inline Javascript code running.
Certainly the xpi file on this site includes "unittest/" folder.